Workaround for Devtools stuck in normal mode
This commit (ab)uses the migrations system to set a non-VimFx pref. The
decision to not inform the user was not taken lightly, but the
preference is vital to being able to type into the devtools, while not
having any drawbacks (this change is part of fission, which is a long
time away and will not cause breakage even if fission were enabled).

Note that the workaround may not be removed from the migrations array,
but only be replaced by an empty function, as a numeric pref will be set
indicating it has been applied.